Financial-agent/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
groq